requirements.txt
aiohttp==3.6.2
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
dhooks==1.1.0
-e git+https://github.com/mobolic/facebook-sdk.git@7a37391940624579cca3d5cdc7370c5f3f83d1f0#egg=facebook_sdk
idna==2.7
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.5.2
pylint==2.4.3
python-dotenv==0.10.3
requests==2.20.1
six==1.13.0
typed-ast==1.4.0
urllib3==1.24.3
wrapt==1.11.2
yarl==1.3.0
